requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
pandas
numpy
openpyxl
